+++ src/main/java/Todo.java
@@ -2,17 +2,17 @@
  * Todo class
  */
 public class Todo extends Task {
-  public Todo (String description) {
-      super(description);
-  }
+    public Todo (String description) {
+        super(description);
+    }
 
-  @Override
-  protected String taskType() {
-      return "T";
-  }
+    @Override
+    protected String taskType() {
+        return "T";
+    }
 
-  @Override
-  public String toString() {
-      return super.toString();
-  }
+    @Override
+    public String toString() {
+        return super.toString();
+    }
 }
